+++ SBTreeFiles/SBBrowserView.m
@@ -74,8 +74,8 @@
 
 -(void)browser:(NSBrowser *)sender 
 		willDisplayCell:(id)cell
-		atRow:(int)row
-		column:(int)column
+		atRow:(NSInteger)row
+		column:(NSInteger)column
 {
     NSWorkspace *ws = [NSWorkspace sharedWorkspace];
     SBFileItem *item;
